feat: clippy cleanup, middleware hooks, and log streaming fix
- Fix all clippy warnings across twerk-app, twerk-web, twerk-infrastructure
- Implement onReadJob/onReadTask middleware hooks for secret masking
- Fix Podman log streaming to initiate before health probes
- Add error docs to public async functions
- Style fixes: match->matches!, inline format args, must_use attrs

+++ crates/twerk-infrastructure/src/runtime/docker/archive.rs
@@ -162,10 +162,9 @@ impl Read for Archive {
 /// Returns the file type bits for the given path.
 ///
 /// This is equivalent to Go's `FileType` function which uses `os.Lstat`.
-#[must_use]
 pub fn file_type(path: &Path) -> Result<u32, ArchiveError> {
     let meta = fs::symlink_metadata(path).map_err(ArchiveError::Io)?;
-    Ok(meta.mode() & S_IFMT as u32)
+    Ok(meta.mode() & S_IFMT)
 }
 
 /// Builder for creating archive entries.
